Module Tjr_path_resolution__Intf.Simplified_result

The result of path resolution (simplified to reduce complexity)

type ('file_id, 'dir_id) simplified_result' =

| File of 'file_id

| Dir of 'dir_id

| Sym of string

| Missing

A "successful" resolution gives one of the following

type ('file_id, 'dir_id) simplified_result = {

parent_id : 'dir_id;

comp : comp_;

result : ('file_id, 'dir_id) simplified_result';

trailing_slash : bool;

}

We also need additional information about the resolved result, such as the containing directory, the last path component, and whether there was a trailing slash on the path